requirements.txt
google-cloud-storage
gcloud
numpy
oauth2client
pandas
tensorflow
matplotlib
httplib2==0.15.0
google-api-python-client==1.6